+++ Sources/VeePreferences/SecretStore.swift
@@ -22,7 +22,7 @@ public struct KeychainSecretStore: SecretStoring {
         [
             kSecClass as String: kSecClassGenericPassword,
             kSecAttrService as String: service,
-            kSecAttrAccount as String: account,
+            kSecAttrAccount as String: account
         ]
     }